exceptions

+++ src/main/java/com/stackroute/MuzixCrudOperations/service/TrackServiceImpl.java
@@ -2,6 +2,7 @@ package com.stackroute.MuzixCrudOperations.service;
 
 import com.stackroute.MuzixCrudOperations.domain.Track;
 import com.stackroute.MuzixCrudOperations.expections.TrackAlreadyExistsException;
+import com.stackroute.MuzixCrudOperations.expections.TrackNotFoundException;
 import com.stackroute.MuzixCrudOperations.repository.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,12 +22,22 @@ public class TrackServiceImpl implements TrackService
     @Override
     public Track saveTrack(Track track) throws TrackAlreadyExistsException
     {
+        if (trackRepository.existsById(track.getId()))
+        {
+            throw new TrackAlreadyExistsException("Track already exists");
+        }
         Track savedTrack=trackRepository.save(track);
+
         return savedTrack;
     }
 
     @Override
-    public void deleteTrack(int id) {
+    public void deleteTrack(int id) throws TrackNotFoundException
+    {
+        if(!trackRepository.findById(id).isPresent())
+        {
+            throw new TrackNotFoundException("No track found with this id");
+        }
         trackRepository.deleteById(id);
 
     }
@@ -37,7 +48,11 @@ public class TrackServiceImpl implements TrackService
     }
 
     @Override
-    public Track getTrackById(int id) {
+    public Track getTrackById(int id) throws TrackNotFoundException {
+        if(!trackRepository.findById(id).isPresent())
+        {
+            throw new TrackNotFoundException("No track is found by this id");
+        }
         Track track=trackRepository.findById(id).get();
         return track;
     }
@@ -48,7 +63,13 @@ public class TrackServiceImpl implements TrackService
         return savedTrack;
     }
     @Override
-    public List<Track> getTrackByName(String name) {
-        return trackRepository.getTrackByName(name);
+    public List<Track> getTrackByName(String name) throws TrackNotFoundException
+    {
+        List<Track> tracksByName = trackRepository.getTrackByName(name);
+        if(tracksByName.isEmpty())
+        {
+            throw new TrackNotFoundException("no track is found by this name");
+        }
+        return tracksByName;
     }
 }
